Added list of tags and installed eslint
