+++ next.config.js
@@ -1,6 +1,6 @@
-const { PHASE_DEVELOPMENT_SERVER } = require('next/constants')
- 
-module.exports = (phase, { defaultConfig }) => {
+const { PHASE_DEVELOPMENT_SERVER } = require('next/constants');
+
+module.exports = (phase) => {
   if (phase === PHASE_DEVELOPMENT_SERVER) {
     return {
       reactStrictMode: true,
@@ -10,17 +10,30 @@ module.exports = (phase, { defaultConfig }) => {
         MONGODB_CLUSTERNAME: 'groupc',
         MONGODB_DATABASE: 'dev-dummy',
       },
-    }
+    };
   }
-
 
   return {
-      reactStrictMode: true,
-  env: {
-    MONGODB_USERNAME: 'groupc',
-    MONGODB_PASSWORD: 'UFX6A2FjZLMhDuME',
-    MONGODB_CLUSTERNAME: 'groupc',
-    MONGODB_DATABASE: 'devdb',
+    reactStrictMode: true,
+    env: {
+      MONGODB_USERNAME: 'groupc',
+      MONGODB_PASSWORD: 'UFX6A2FjZLMhDuME',
+      MONGODB_CLUSTERNAME: 'groupc',
+      MONGODB_DATABASE: 'devdb',
+    },
+  };
+};
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+
+  images: {
+    formats: ['image/avif', 'image/webp'],
+    domains: ['img.sndimg.com', 'images.pexels.com'], // Updated from remotePatterns to domains
   },
-  }
-}
+  eslint: {
+    dirs: ['pages/tags'],
+  },
+
+};
