+++ package.json
@@ -27,6 +27,7 @@
   "devDependencies": {
     "autoprefixer": "^10",
     "eslint": "^8",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-next": "13.5.4",
     "postcss": "^8",
     "tailwindcss": "^3.3.3"
